feat: add OrderCreate and OrderOut schemas; update AdminAuthService and UserAuthService to use Member model

+++ Timur_Medikhanov_Task_7/api/products/schemas.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 from datetime import datetime
 
@@ -5,7 +7,7 @@ from datetime import datetime
 class ProductIn(BaseModel):
     name: str
     description: str | None = None
-    price: int
+    price: Decimal
 
 
 class ProductOut(ProductIn):
